permanent_makeup/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==4.2.7
django-bootstrap-datepicker-plus==5.0.4
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
sqlparse==0.4.4
typing_extensions==4.8.0
